refactor（v0.2.1-optimal）：目录化 messages_to_responses direction
